rename BasicAnalysis -> Analysis and define alias for backwards compatibility, move mcmc_sampling to own module (shouldn't and doesn't need to be part of Analysis class)
