+++ pisa/analysis/analysis.py
@@ -1,6 +1,6 @@
 """
 Common tools for performing an analysis collected into the class
-`BasicAnalysis` that can be subclassed by specific analyses.
+`Analysis` that can be subclassed by specific analyses.
 """
 
 
@@ -11,7 +11,6 @@ from operator import setitem
 import re
 import sys
 import time
-import warnings
 
 import numpy as np
 import scipy
@@ -24,7 +23,7 @@ from iminuit import Minuit
 import nlopt
 from pkg_resources import parse_version
 
-from pisa import FTYPE, ureg
+from pisa import ureg
 from pisa.analysis.configure_nlopt_minimization import (
     get_nlopt_inequality_constraint_funcs
 )
@@ -44,13 +43,12 @@ from pisa.utils.comparisons import recursiveEquality, FTYPE_PREC
 from pisa.utils.config_parser import parse_pipeline_config
 from pisa.utils.log import logging, set_verbosity
 from pisa.utils.fileio import EVAL_MSG, from_file
-from pisa.utils.random_numbers import get_random_state
 from pisa.utils.stats import (METRICS_TO_MAXIMIZE, METRICS_TO_MINIMIZE,
                               LLH_METRICS, CHI2_METRICS, weighted_chi2,
                               it_got_better, is_metric_to_maximize)
 
 __all__ = ['SUPPORTED_LOCAL_SCIPY_MINIMIZERS', 'MINIMIZERS_USING_SYMM_GRAD',
-           'MINIMIZERS_ACCEPTING_CONSTRS', 'BasicAnalysis', 'Counter',
+           'MINIMIZERS_ACCEPTING_CONSTRS', 'Analysis', 'BasicAnalysis', 'Counter',
            'HypoFitResult']
 
 __author__ = 'J.L. Lanfranchi, P. Eller, S. Wren, E. Bourbeau, A. Trettin, T. Ehrhardt'
@@ -483,11 +481,12 @@ class HypoFitResult():
         return detailed_metric_info
 
 
-class BasicAnalysis():
-    """A bare-bones analysis that only fits a hypothesis to data.
+class Analysis():
+    """An analysis class that performs a flexible (global, local, nested, ...) fit of
+    a hypothesis to data.
 
     Full analyses with functionality beyond just fitting (doing scans, for example)
-    should sub-class this class.
+    are outside the scope of this class and should be defined by the user.
 
     Every fit is run with the `fit_recursively` method, where the fit strategy is
     defined by the three arguments `method`, `method_kwargs` and
@@ -703,10 +702,10 @@ class BasicAnalysis():
     `external_priors_penalty`, `method_kwargs`, and `local_fit_kwargs`. See docstring
     of `fit_recursively` for descriptions of these arguments. The return value
     of the function must be a `HypoFitResult` object. As an example, the following
-    sub-class of the BasicAnalysis has a custom fit method that, nonsensically,
+    sub-class of `Analysis` has a custom fit method that, nonsensically,
     always sets 42 degrees as the starting value for theta23:
     ::
-        class SubclassedAnalysis(BasicAnalysis):
+        class SubclassedAnalysis(Analysis):
 
             def _fit_nonsense(
                 self, data_dist, hypo_maker, metric,
@@ -2615,143 +2614,19 @@ class BasicAnalysis():
         """
         self._nit += 1
 
-    def MCMC_sampling(self, data_dist, hypo_maker, metric, nwalkers, burnin, nsteps,
-                      return_burn_in=False, random_state=None, sampling_algorithm=None):
-        """Performs MCMC sampling. Only supports serial (single CPU) execution at the
-        moment. See issue #830.
 
-        Parameters
-        ----------
-
-        data_dist : Sequence of MapSets or MapSet
-            Data distribution to be fit. Can be an actual-, Asimov-, or pseudo-data
-            distribution (where the latter two are derived from simulation and so
-            aren't technically "data").
-
-        hypo_maker : Detectors or DistributionMaker
-            Creates the per-bin expectation values per map based on its param values.
-            Free params in the `hypo_maker` are modified by the minimizer to achieve a
-            "best" fit.
-
-        metric : string or iterable of strings
-            Metric by which to evaluate the fit. See documentation of Map.
-
-        nwalkers : int
-            Number of walkers
-
-        burnin : int
-            Number of steps in burn in phase
-
-        nSteps : int
-            Number of steps after burn in
-
-        return_burn_in : bool
-            Also return the steps of the burn in phase. Default is False.
-
-        random_state : None or type accepted by utils.random_numbers.get_random_state
-            Random state of the walker starting points. Default is None.
-
-        sampling_algorithm : None or emcee.moves object
-            Sampling algorithm used by the emcee sampler. None means to use the default
-            which is a Goodman & Weare “stretch move” with parallelization.
-            See https://emcee.readthedocs.io/en/stable/user/moves/#moves-user to learn
-            more about the emcee sampling algorithms.
-
-        Returns
-        -------
-
-        scaled_chain : numpy array
-            Array containing all points in the parameter space visited by each walker.
-            It is sorted by steps, so all the first steps of all walkers come first.
-            To for example get all values of the Nth parameter and the ith walker, use
-            scaled_chain[i::nwalkers, N].
-
-        scaled_chain_burnin : numpy array (optional)
-            Same as scaled_chain, but for the burn in phase.
-
-        """
-        import emcee # pylint: disable=import-outside-toplevel
-
-        assert 'llh' in metric or 'chi2' in metric, 'Use either a llh or chi2 metric'
-        if 'chi2' in metric:
-            warnings.warn("You are using a chi2 metric for the MCMC sampling."
-                          "The sampler will assume that llh=0.5*chi2.")
-
-        ndim = len(hypo_maker.params.free)
-        bounds = np.repeat([[0,1]], ndim, axis=0)
-        rs = get_random_state(random_state)
-        p0 = rs.random(ndim * nwalkers).reshape((nwalkers, ndim))
-
-        def func(scaled_param_vals, bounds, data_dist, hypo_maker, metric):
-            """Function called by the MCMC sampler. Similar to _minimizer_callable it
-            returns the current metric value + prior penalties.
-
-            """
-            if (np.any(scaled_param_vals > np.array(bounds)[:, 1]) or
-                np.any(scaled_param_vals < np.array(bounds)[:, 0])):
-                return -np.inf
-            sign = +1 if metric in METRICS_TO_MAXIMIZE else -1
-            if 'llh' in metric:
-                N = 1
-            elif 'chi2' in metric:
-                N = 0.5
-
-            hypo_maker._set_rescaled_free_params(scaled_param_vals) # pylint: disable=protected-access
-            hypo_asimov_dist = hypo_maker.get_outputs(return_sum=True)
-            metric_val = (
-                N * data_dist.metric_total( # pylint: disable=possibly-used-before-assignment
-                    expected_values=hypo_asimov_dist, metric=metric)
-                + hypo_maker.params.priors_penalty(metric=metric)
-            )
-            return sign*metric_val
-
-        sampler = emcee.EnsembleSampler(
-            nwalkers, ndim, func,
-            moves=sampling_algorithm,
-            args=[bounds, data_dist, hypo_maker, metric]
-        )
-
-        if self.pprint:
-            sys.stdout.write('Burn in')
-            sys.stdout.flush()
-        pos, prob, state = sampler.run_mcmc(p0, burnin, progress=self.pprint) # pylint: disable=unused-variable
-
-        if return_burn_in:
-            flatchain_burnin = sampler.flatchain
-            scaled_chain_burnin = np.full_like(flatchain_burnin, np.nan, dtype=FTYPE)
-            param_copy_burnin = ParamSet(hypo_maker.params.free)
-
-            for s, sample in enumerate(flatchain_burnin):
-                for dim, rescaled_val in enumerate(sample):
-                    param = param_copy_burnin[dim]
-                    param._rescaled_value = rescaled_val # pylint: disable=protected-access
-                    val = param.value.m
-                    scaled_chain_burnin[s, dim] = val
-
-        sampler.reset()
-        if self.pprint:
-            sys.stdout.write('Main sampling')
-            sys.stdout.flush()
-        sampler.run_mcmc(pos, nsteps, progress=self.pprint)
-
-        flatchain = sampler.flatchain
-        scaled_chain = np.full_like(flatchain, np.nan, dtype=FTYPE)
-        param_copy = ParamSet(hypo_maker.params.free)
-
-        for s, sample in enumerate(flatchain):
-            for dim, rescaled_val in enumerate(sample):
-                param = param_copy[dim]
-                param._rescaled_value = rescaled_val # pylint: disable=protected-access
-                val = param.value.m
-                scaled_chain[s, dim] = val
-
-        if return_burn_in:
-            return scaled_chain, scaled_chain_burnin
-        return scaled_chain
+# provide an alias for backwards compatibility
+BasicAnalysis = Analysis
+"""
+>> BasicAnalysis.__name__
+"Analysis"
+>> BasicAnalysis.__qualname__
+"Analysis"
+"""
 
 
-def test_basic_analysis(pprint=False):
-    """Test recursive fit strategies with BasicAnalysis."""
+def test_analysis(pprint=False):
+    """Test recursive fit strategies with Analysis."""
 
     ###### Make Pipeline Configuration #########
     #  We make a configuration of two pipelines where some, but not all, parameters
@@ -2771,9 +2646,9 @@ def test_basic_analysis(pprint=False):
     )
 
     #### Test subclassing
-    # It should be trivial to add a fit method to the BasicAnalysis class and use
+    # It should be trivial to add a fit method to the Analysis class and use
     # it by passing its name (without the "_fit_" prefix) to the dictionary.
-    class SubclassedAnalysis(BasicAnalysis):
+    class SubclassedAnalysis(Analysis):
 
         def _fit_nonsense(
             self, data_dist, hypo_maker, metric,
@@ -3077,7 +2952,7 @@ def test_basic_analysis(pprint=False):
     assert best_fit_info.fit_history[0][0] == "chi2"
     assert isinstance(best_fit_info.fit_history[1][0], float)
 
-    logging.info('<< PASS : test_basic_analysis >>')
+    logging.info('<< PASS : test_analysis >>')
 
 
 def test_constrained_minimization(pprint=False):
@@ -3091,7 +2966,7 @@ def test_constrained_minimization(pprint=False):
 
     ### slsqp test with constraints ###
     def slsqp_constr():
-        ana = BasicAnalysis()
+        ana = Analysis()
         ana.pprint = pprint
         min_sett = {
           "method": {"value": "slsqp", "desc": ""},
@@ -3138,7 +3013,7 @@ def test_constrained_minimization(pprint=False):
 
     ### cobyla test with inequality constraints (doesn't support equalities) ###
     def cobyla_constr():
-        ana = BasicAnalysis()
+        ana = Analysis()
         ana.pprint = pprint
 
         min_t23 = 46.
@@ -3180,7 +3055,7 @@ def test_constrained_minimization(pprint=False):
 
     ### trust-constr test with constraints ###
     def trust_constr_constr():
-        ana = BasicAnalysis()
+        ana = Analysis()
         ana.pprint = pprint
 
         min_delta_index = 5e-3
@@ -3225,7 +3100,7 @@ def test_constrained_minimization(pprint=False):
 
     ### Nelder-Mead test (no constraints supported) ###
     def nm_unconstr():
-        ana = BasicAnalysis()
+        ana = Analysis()
         ana.pprint = pprint
 
         min_sett = {
@@ -3253,7 +3128,7 @@ def test_constrained_minimization(pprint=False):
 
     ### finally an nlopt solver with constraints ###
     def some_nlopt_constr():
-        ana = BasicAnalysis()
+        ana = Analysis()
         ana.pprint = pprint
 
         min_t23 = 46.
@@ -3318,7 +3193,7 @@ def global_scipy_minimization(pprint=False):
     def run_global_with_supported_local(
             global_method_kwargs, constrs_list=None, constrs_test=None
     ):
-        ana = BasicAnalysis()
+        ana = Analysis()
         ana.pprint = pprint
 
         assert (constrs_list is None) == (constrs_test is None)
@@ -3364,7 +3239,7 @@ def global_scipy_minimization(pprint=False):
 
 
     def run_differential_evolution(constrs_list, constrs_test, polish=True):
-        ana = BasicAnalysis()
+        ana = Analysis()
         ana.pprint = pprint
 
         assert (constrs_list is None) == (constrs_test is None)
@@ -3453,7 +3328,7 @@ def global_scipy_minimization(pprint=False):
 
     def run_global_min_with_constraints_from_file(fit_sett):
 
-        ana = BasicAnalysis()
+        ana = Analysis()
         ana.pprint = pprint
 
         fit_sett = from_file(fit_sett)
@@ -3481,7 +3356,7 @@ def global_scipy_minimization(pprint=False):
 
 if __name__ == "__main__":
     set_verbosity(1)
-    test_basic_analysis(pprint=True)
+    test_analysis(pprint=True)
     test_constrained_minimization(pprint=True)
     #TODO: rename to test_global_scipy_minimization in case auto detection re-enabled (see above)
     global_scipy_minimization(pprint=True)
